Use size_t for "size".

+++ testsuites/sptests/sp09/screen11.c
@@ -24,7 +24,7 @@ void Screen11()
   void              *buffer_address_2;
   void              *buffer_address_3;
   rtems_status_code  status;
-  uint32_t           size;
+  size_t             size;
 
   status = rtems_partition_create(
     0,
@@ -106,7 +106,7 @@ void Screen11()
       &Junk_id
     );
     if ( status != RTEMS_INVALID_SIZE )
-      printf( "ERROR when size == %d\n", size );
+      printf( "ERROR when size == %zu\n", size );
 
     fatal_directive_status(
       status,
